Updated TS_groups for intra_H_mig and R_Add_MultiBond
R_Add_MultipleBond: Updated the group tree and retrained the
data.

intra_H_migration: The reverse reactions in the training 
data were duplicated from the forward reactions incorrectly.
As a result, the groups were improperly trained. This was
corrected, with the groups being retrained with the proper
reverse reactions. 4 more transition states were also added
to the training data.