Convert TS_training.py to new training set style
The kinetics training and NIST data were converted to
folders containing `reactions.py` and `dictionary.txt`. The
transition state training data was converted to take advantage
of the benefits of the change as well as for database consistency.
`reactions.py` contains reaction entries along with transition state
data. The entries no longer have reactant and product adjacency lists,
but instead use the `entry.label` to sore the reactant and product keys.
A given adjacency list is sourced from `dictionary.txt`. Methods to work
with the new database style are available in `rmgpy.data.kinetics`