Corrected NN_ads bond type to Pt surface
Changed to vdW bond in adjacency lists of surfacethermoPt and adsorptionPt file.
Fixed a slight difference that makes in the thermo and adsorption corrections for adsorbed NN.
Fixed linear scaling comments.
Commented out NN entry in adsorptionPt file, since it is not needed, and removed N-*N-* node in tree.