Restrict partial charge in Surface_Adsorption_Abstraction_vdW
We don't want to break a bond between two charged atoms,
eg. [O-]---[N+](=O)O 
because then you'd have charged adsorbates.

Maybe we *do* want to be able to break such bonds, but
then do something with the charges.
But hopefully this way we'll just find a different resonance
structure and break that bond instead.