Don't apply 1+2_Cycloaddition to create Halogens with 3 bonds.
The recipe takes the *3 atom which has at least 1 lone pair
(and a bond) and turns the lone pair into two more bonds, 
so the atom ends up with 3 bonds. 
This doesn't make sense for a Val7 atom (valence 7) which 
don't want to have 3 bonds. It would lead to an AtomTypeError
which would probably be handled by not generating the product,
but wastes time and creates logging.

Here we make sure *3 is R!H!Val7 instead of R.

Possibly the tree should be re-trained.